Moved trainer slides to their own file Moved trainer slide check conditions to their own functions
